inference/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
tensorflow>=2.16.0
numpy==1.26.3
Pillow==10.2.0
h5py>=3.10.0
python-dotenv==1.0.1
